requirements.txt
discord.py>=2.0.0
langdetect>=1.0.9